Better class Runner.
