+++ src/runner.ts
@@ -1,0 +1,53 @@
+import { afterAll, describe, it } from "bun:test";
+import parser from "@typescript-eslint/parser";
+import {
+  type InvalidTestCase,
+  RuleTester,
+  type ValidTestCase,
+} from "@typescript-eslint/rule-tester";
+import type { RuleModule } from "@typescript-eslint/utils/ts-eslint";
+
+export class Runner<
+  Options extends readonly unknown[],
+  MessageId extends string,
+> {
+  constructor(
+    protected title: string,
+    protected ruleModule: RuleModule<MessageId, Options>,
+    protected scenarios: Record<
+      string,
+      (ValidTestCase<Options> | InvalidTestCase<MessageId, Options>) & {
+        before?: () => void;
+      }
+    >,
+  ) {
+    RuleTester.afterAll = afterAll;
+    RuleTester.describe = describe;
+    RuleTester.it = (...[name, ...rest]: Parameters<typeof it>) => {
+      const { before } = scenarios[name];
+      before?.();
+      it(name, ...rest);
+    };
+  }
+
+  public run() {
+    const tester = new RuleTester({
+      languageOptions: { parser },
+    });
+    const [valid, invalid] = Object.keys(this.scenarios).reduce<
+      [ValidTestCase<Options>[], InvalidTestCase<MessageId, Options>[]]
+    >(
+      ([a, b], name) => {
+        const { before, ...scenario } = this.scenarios[name];
+        const hasErrors = "errors" in scenario;
+        (hasErrors ? b : a).push({
+          name,
+          ...scenario,
+        });
+        return [a, b];
+      },
+      [[], []],
+    );
+    tester.run("dependencies", this.ruleModule, { valid, invalid });
+  }
+}
